Hotfix - converted type variables Fixed error: converting driver.Value type []uint8 ("110680464442257309690") to a int: value out of range

+++ models/models.go
@@ -29,8 +29,8 @@ type Metrics struct {
 	DB struct {
 		Metrics struct {
 			Status                               MetricGroupValue
-			TotalTables                          int
-			TotalMyisamIndexes                   int
+			TotalTables                          uint64
+			TotalMyisamIndexes                   uint64
 			Engine                               map[string]MetricGroupValue
 			Latency                              string
 			QueriesLatency                       []MetricGroupValue
